* add 'company_id' to User model * move User model to app/Models, update references * add migration, factory and seeder for User model
